+++ src/com/github/tnerevival/core/configurations/impl/MaterialsConfiguration.java
@@ -14,11 +14,11 @@ public class MaterialsConfiguration extends Configuration {
 	
 	@Override
 	public void load(FileConfiguration configurationFile) {
+    Set<String> identifiers = TNE.instance.worldConfigurations.getConfigurationSection("Worlds").getKeys(false);
+
 		//Load Materials
     loadMaterials(configurationFile, "", null, true);
     loadMaterials(configurationFile, "", null, false);
-
-    Set<String> identifiers = TNE.instance.worldConfigurations.getConfigurationSection("Worlds").getKeys(false);
     for(String identifier : identifiers) {
       loadMaterials(TNE.instance.worldConfigurations, "Worlds." + identifier + ".", identifier, true);
       loadMaterials(TNE.instance.worldConfigurations, "Worlds." + identifier + ".", identifier, false);
@@ -29,14 +29,14 @@ public class MaterialsConfiguration extends Configuration {
       loadMaterials(TNE.instance.playerConfigurations, "Players." + identifier + ".", identifier, true);
       loadMaterials(TNE.instance.playerConfigurations, "Players." + identifier + ".", identifier, false);
     }
-
 		super.load(configurationFile);
 	}
 
 	private void loadMaterials(FileConfiguration configuration, String baseNode, String identifier, boolean item) {
 	  String base = baseNode + ((item)? "Materials.Items" : "Materials.Blocks");
 	  if(configuration.contains(base)) {
-      configurations.put(baseNode + ".ZeroMessage", false);
+	    Boolean zero = (configuration.contains(base + ".ZeroMessage"))? configuration.getBoolean(base + ".ZeroMessage") : true;
+      configurations.put(base + ".ZeroMessage", zero);
 
       Set<String> materialNames = configuration.getConfigurationSection(base).getKeys(false);